refactor(docker): reorganize Docker output classes into a dedicated namespace
Move Docker output-related classes into a new `Output` namespace to improve
code organization and maintainability. This change enhances the logical
grouping of classes related to Docker command outputs, making it easier
to manage and extend. Additionally, it introduces new classes for handling
specific Docker command outputs, such as `DockerFollowLogsOutput` and
`DockerInspectOutput`, providing more structured and reusable code for
processing Docker command results.
