--- src/Docker/DockerNetworkCreateOutput.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Testcontainers\Docker;
-
-class DockerNetworkCreateOutput extends DockerOutput
-{
-}
