+++ src/Docker/Output/DockerFollowLogsOutput.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Testcontainers\Docker;
+namespace Testcontainers\Docker\Output;
 
 use Generator;
+use Testcontainers\Docker\Output\DockerOutput;
 
 /**
  * Represents the output of a Docker `logs` command executed via Symfony Process.
