+++ src/Docker/Output/DockerNetworkCreateOutput.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Testcontainers\Docker\Output;
+
+use Testcontainers\Docker\Output\DockerOutput;
+
+class DockerNetworkCreateOutput extends DockerOutput
+{
+}
