+++ src/Docker/Output/DockerOutput.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Testcontainers\Docker;
+namespace Testcontainers\Docker\Output;
 
 use Symfony\Component\Process\Process;
