+++ src/Docker/Output/DockerProcessStatusOutput.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Testcontainers\Docker;
+namespace Testcontainers\Docker\Output;
 
 use LogicException;
 use Symfony\Component\Process\Process;
+use Testcontainers\Docker\Output\DockerOutput;
 
 /**
  * Represents the status output of a Docker process.
