+++ src/Docker/Output/DockerRunOutput.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Testcontainers\Docker;
+namespace Testcontainers\Docker\Output;
+
+use Testcontainers\Docker\Output\DockerOutput;
 
 /**
  * Represents the output of a Docker `run` command executed via Symfony Process.
