+++ tests/Unit/Docker/Command/RunCommandTest.php
@@ -5,8 +5,8 @@ namespace Tests\Unit\Docker\Command;
 use PHPUnit\Framework\TestCase;
 use Testcontainers\Docker\Command\RunCommand;
 use Testcontainers\Docker\DockerClient;
-use Testcontainers\Docker\DockerRunOutput;
-use Testcontainers\Docker\DockerRunWithDetachOutput;
+use Testcontainers\Docker\Output\DockerRunOutput;
+use Testcontainers\Docker\Output\DockerRunWithDetachOutput;
 use Testcontainers\Docker\Exception\PortAlreadyAllocatedException;
 use Testcontainers\Testcontainers;
 use Tests\Images\DinD;
